Move most types to jetstream_types module
